Credit_Scoring_MLOps/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
onnx
onnxruntime
skl2onnx
onnxconverter_common
evidently
prometheus-fastapi-instrumentator
dvc[ssh]
apache-airflow
apache-airflow-providers-cncf-kubernetes
pyyaml
safety
redis
